feat: consolidate donuts in one total stats component
