--- src/app/india-elections-2024/components/total-seats/total-seats.component.html
@@ -1,3 +1,0 @@
-<app-donut-chart (sectorMouseover)="onSectorMouseover($event)" (sectorMouseout)="onSectorMouseout()"
-                 [data]="seatsData()" [valueFn]="valueFn" [labelFn]="labelFn" [colorFn]="colorFn"
-                 title="Total seats won"/>
