+++ src/app/india-elections-2024/components/total-stats/total-stats.component.scss
@@ -1,5 +1,9 @@
 :host {
   display: flex;
   flex-direction: column;
-  align-items: center;
+  gap: 1rem;
+
+  .charts {
+    max-width: 150px;
+  }
 }
