--- src/app/india-elections-2024/components/total-vote-share/total-vote-share.component.html
@@ -1,3 +1,0 @@
-<app-donut-chart (sectorMouseover)="onSectorMouseover($event)" (sectorMouseout)="onSectorMouseout()"
-                 [data]="votePctData()" [valueFn]="valueFn" [labelFn]="labelFn" [colorFn]="colorFn"
-                 title="Total vote share" />
