--- src/app/india-elections-2024/components/total-vote-share/total-vote-share.component.scss
@@ -1,5 +1,0 @@
-:host {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
