+++ src/app/india-elections-2024/india-elections-2024.component.html
@@ -1,10 +1,7 @@
 <h4>2024 Indian Parliamentary Election Results</h4>
 <section class="map">
   <app-constituencies-map (constituencyClick)="onConstituencyClick($event)" [highlight]="highlightConstituencies()"/>
-  <div class="total-stats">
-    <app-total-seats (partyHover)="onPartyHover($event)"/>
-    <app-total-vote-share (partyHover)="onPartyHover($event)"/>
-  </div>
+  <app-total-stats (partyHover)="onPartyHover($event)" />
 </section>
 <section class="constituency-detail">
   <app-constituency-detail [constituency]="selectedConstituency()"/>
